+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/builder/impl/SearchWorkBuilder.java
@@ -31,5 +31,5 @@ public interface SearchWorkBuilder<R> extends ElasticsearchWorkBuilder<NonBulkab
 
 	SearchWorkBuilder<R> disableTrackTotalHits();
 
-	SearchWorkBuilder<R> totalHitCountThreshold(Integer totalHitCountThreshold);
+	SearchWorkBuilder<R> totalHitCountThreshold(Long totalHitCountThreshold);
 }
