+++ engine/src/main/java/org/hibernate/search/engine/search/query/dsl/spi/AbstractSearchQueryOptionsStep.java
@@ -107,7 +107,7 @@ public abstract class AbstractSearchQueryOptionsStep<
 	}
 
 	@Override
-	public S totalHitCountThreshold(int totalHitCountThreshold) {
+	public S totalHitCountThreshold(long totalHitCountThreshold) {
 		searchQueryBuilder.totalHitCountThreshold( totalHitCountThreshold );
 		return thisAsS();
 	}
